requirements.txt
aiomysql==0.2.0
aiosqlite==0.17.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
autopep8==2.0.2
bcrypt==4.0.1
bidict==0.23.1
bleach==6.2.0
blinker==1.8.2
boto3==1.35.54
botocore==1.35.54
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
cs50==9.3.4
databases==0.9.0
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
emoji==2.4.0
fastapi==0.115.5
fastapi-cli==0.0.5
fastapi-users==14.0.0
filelock==3.16.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-SocketIO==5.3.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.9
iniconfig==2.0.0
iso8601==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jose==1.0.0
makefun==1.15.6
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysql-connector-python==9.0.0
orjson==3.10.7
packaging==24.1
passlib==1.7.4
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.9
pwdlib==0.2.1
pyasn1==0.6.1
pycodestyle==2.10.0
pycparser==2.22
pydantic==2.10.0
pydantic-extra-types==2.9.0
pydantic-settings==2.6.1
pydantic_core==2.27.0
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.8.0
PyMySQL==1.1.1
pypika-tortoise==0.1.6
pytest==8.3.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-engineio==4.9.1
python-jose==3.3.0
python-multipart==0.0.17
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
rich==13.8.1
rsa==4.9
s3transfer==0.10.3
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
starlette==0.41.3
termcolor==2.4.0
tortoise-orm==0.20.0
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.22.0
virtualenv==20.26.5
watchfiles==0.24.0
webencodings==0.5.1
websockets==13.0.1
Werkzeug==3.0.4
wsproto==1.2.0
